Improve build of libr.a and libr.dylib, fix and improve sys/ios-sdk.sh
- Update spp and force hidden visibility
- Use R_API wisely
- RSys.prefix returns const things
- Use -install_name on Apple
- Fix merged lib visibility linking on Linux
- Use OSTYPE instead of BUILD_OS
- Honor crosscompiler-objcopy and support android like linux
- Add extra missing archives
- Fix for android
- Upgrade spp and sdb
- Skip libr. from symstall
- Add --enable-merged configure option
- Use --enable-merged on ios-sdk
- Upgrade sdb again for js0n
- Kill sys/ios-shell.sh

+++ libr/debug/p/native/linux/linux_debug.c
@@ -1,4 +1,4 @@
-/* radare - LGPL - Copyright 2009-2016 - pancake */
+/* radare - LGPL - Copyright 2009-2018 - pancake */
 
 #include <r_userconf.h>
 
@@ -105,7 +105,7 @@ int linux_handle_signals (RDebug *dbg) {
 				}
 				if (dbg->reason.type != R_DEBUG_REASON_NEW_LIB &&
 					dbg->reason.type != R_DEBUG_REASON_EXIT_LIB) {
-					dbg->reason.bp_addr = (ut64)siginfo.si_addr;
+					dbg->reason.bp_addr = (ut64)(size_t)siginfo.si_addr;
 					dbg->reason.type = R_DEBUG_REASON_BREAKPOINT;
 				}
 			}
@@ -758,7 +758,7 @@ void print_fpu (void *f, int r){
 	}
 #endif
 #else
-#warning not implemented for this platform
+#warning print_fpu not implemented for this platform
 #endif
 }
 
@@ -853,7 +853,7 @@ int linux_reg_read (RDebug *dbg, int type, ut8 *buf, int size) {
 		}
 		}
 #else
-	#warning not implemented for this platform
+	#warning getfpregs not implemented for this platform
 #endif
 		break;
 	case R_REG_TYPE_SEG:
